+++ nutella/@types/emotion/index.d.ts
@@ -1,5 +1,5 @@
 import "@emotion/react";
-import { Theme as ThemeType } from "../../src/style/theme";
+import { Theme as ThemeType } from "../../src/utils/theme/theme";
 
 declare module "@emotion/react" {
   export interface Theme extends ThemeType {}
